Time: 4 ms (42.20%) | Memory: 69.8 MB (10.54%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.java
@@ -11,6 +11,6 @@ class Solution {
 
     max = Math.max(max, sum);
         }
-        return max/k;
+        return (double)max/k;
     }
 }
